Moved: components pickers to pickers folder
