+++ apps/frontend/src/components/page-products/pickers/PickerAttributes.vue
@@ -1,5 +1,5 @@
 <script>
-import ItemAttrValue from '@/components/page-products/ItemAttrValue.vue'
+import ItemAttrValue from '~/page-products/ItemAttrValue.vue'
 
 export default {
   components: { ItemAttrValue },
