+++ apps/frontend/src/components/page-products/pickers/PickerCurrentPage.vue
