+++ apps/frontend/src/components/page-products/pickers/PickerListType.vue
