+++ apps/frontend/src/components/page-products/pickers/PickerPageSize.vue
