+++ apps/frontend/src/components/page-products/pickers/PickerPriceRanges.vue
