+++ apps/frontend/src/components/page-products/pickers/PickerSearchQuery.vue
