+++ apps/frontend/src/components/page-products/pickers/PickerSortingType.vue
